automod: quick extraction of redis directory cache.
This one will probably move further in the future -- there are already
directory service mocks elsewhere, for example, and so this should most
likely live near those.  That may deserve a larger code review,
however, so for now we just move it into a package as we do with other
things today.
